requirements.txt
Django==2.2.13
djangorestframework==3.11.0

django-filter==2.3.0
python-dotenv==0.13.0
pymongo~=3.10.1
pandas~=1.0.5
psycopg2-binary==2.8.5
xlrd==1.2.0